feat: code split plotly pt 2
