+++ src/components/my-career/data/hooks/index.js
@@ -1,1 +1,3 @@
+export { default as useLearnerProfileData } from './useLearnerProfileData';
+export { default as useLearnerSkillLevels } from './useLearnerSkillLevels';
 export { default as usePlotlySpiderChart } from './usePlotlySpiderChart';
